+++ crates/kreuzberg/src/pdf/rendering.rs
@@ -68,7 +68,7 @@ impl PdfRenderer {
 
         let page = document
             .pages()
-            .get(page_index as i32)
+            .get(page_index as u16)
             .map_err(|_| PdfError::PageNotFound(page_index))?;
 
         let width_points = page.width().value;
